remove Steps and refactor its test with BasicStepFactory
