+++ src/test/java/nwoolcan/model/brewery/batch/step/BasicStepFactoryTest.java
@@ -1,9 +1,5 @@
-package nwoolcan.model.brewery.batch.step.utils;
+package nwoolcan.model.brewery.batch.step;
 
-import nwoolcan.model.brewery.batch.step.Step;
-import nwoolcan.model.brewery.batch.step.StepType;
-import nwoolcan.model.brewery.batch.step.StepTypeEnum;
-import nwoolcan.model.brewery.batch.step.Steps;
 import nwoolcan.utils.Result;
 
 import org.junit.Assert;
@@ -12,21 +8,23 @@ import org.junit.Test;
 import java.util.Date;
 
 /**
- * Test class for Steps util.
+ * Test class for factory util.
  */
-public class StepsTest {
+public class BasicStepFactoryTest {
+
+    private final StepFactory factory = new BasicStepFactory();
 
     /**
      * Test simple step creation.
      */
     @Test
     public void testSimpleCreation() {
-        Result<Step> res = Steps.create(StepTypeEnum.AGING);
+        Result<Step> res = factory.create(StepTypeEnum.AGING);
         Assert.assertTrue(res.isPresent());
         Assert.assertEquals(StepTypeEnum.AGING, res.getValue().getStepInfo().getType());
 
         Date d = new Date();
-        res = Steps.create(StepTypeEnum.MASHING, d);
+        res = factory.create(StepTypeEnum.MASHING, d);
         Assert.assertTrue(res.isPresent());
         Assert.assertEquals(StepTypeEnum.MASHING, res.getValue().getStepInfo().getType());
         Assert.assertEquals(d, res.getValue().getStepInfo().getStartDate());
@@ -37,7 +35,7 @@ public class StepsTest {
      */
     @Test
     public void testWrongCreation() {
-        Result<Step> res = Steps.create(new StepType() {
+        Result<Step> res = factory.create(new StepType() {
             @Override
             public String getName() {
                 return "test";
